requirements.txt
Django
requests
djangorestframework
django-mysql
django-bootstrap4
django-octicons
django-widget-tweaks
gunicorn
